requirements.txt
aiogram==3.10.0
SQLAlchemy==2.0.28
aiosqlite
python-dotenv==1.0.0
pandas
openpyxl==3.1.2
psutil==5.9.6